Route any URL to a playground (+prep persistence)
Any non static URL is routed to a playground.
An ID is created based on the URL/
A global object 'all playgrounds' (bad, I was fighting endlessly with
app.locals, require and exports…) is used to check the previous
existence of the playground. If non-existent an object is created for
the playground in 'allplaygrounds'
TODO: handling the arbitrary URL as a proper identifier, i.e. how is
Etherpad doing it ??