rewriting background() calls to always be transparent
Opaque backgrounds were found to be really annoying when testing
Paysage at the Code Week closing event (the last object mask
everything).
Setting it to transparent using the instance method ex:
sketch[].background(0.0) doesn't stick if background() is in draw(),
because it's redrawn each frame!
So we rewrite the code just before creating the processing instance
with it.