Better background defaults
Instead of rewriting the user's code to force a transparent background,
we now monkey-patch Processing.js to set the default value of the alpha
component to zero. This means that solid backgrounds will not appear by
default, unless the alpha component is overridden.

Also, we set the default background color to fully transparent so that a
sketch that never calls background() will be transparent instead of a
dull grey.