fix drag and drop of files, extend to body
ACE editor broke the updating of the editor content when drag and
dropping a text file (.pde or .pjs files for example). Use
Paysage.setCode now to update.
Also, drag and dropping was restricted to the editor text zone, one can
now drop anywhere in the page body. Avoid mistargeting ;-)