Send object list on connect even if empty
The 'objects list'/'playground full update' messages were not sent if
the playground was empty. This caused desynchronization when a
programmer was connected to a non-empty playground, was disconnected
(because the server restarted, for example), then reconnected: the old
objects would linger in the programmer's list until the next object
update.

Likewise for the renderer, which also now knows to clear its running
objects on a full update.