Deactivate firefox autocomplete for the checkbox of the switch preview button
On firefox, if you reload (simple reload without reloading all the resources)
the page while the switch button is on "preview is off", the switch become
desynchronised from the state of the page. It appears grey and on the left,
while the text says "preview is on" and the preview is effectively on.

https://bugzilla.mozilla.org/show_bug.cgi?id=654072