Rename a creature (#110)
In order to avoid renaming the id of code object, I introduced "name" property on CodeObject.
This name is not sent to the renderer. The renderer is not impacted by this PR.

Programmers' view know contains two input fields, name and id. Id is hidden.

The user should enter a name after loading a new code object.